chore: enable uglify compression
- Upgrade the uglifyjs-webpack-plugin to the latest.
- Update prod webpack conf with uglify settings to preserve names.

There is a weird issue that means the running the same
minification steps via a yarn installed dep tree vs an npm one
gives different results. This PR adds ensures that we get some
useful compression on our bundles but ensure that we dont break
IPFS while doing so.

The build now works on npm and yarn, which is nice, and local
testing suggest the IPFS boot time is much faster, which is ✨

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>