feat: add error reporting via rollbar (#227)
* feat: add error reporting via rollbar

Adds reporting of errors when they happen on peerpad.net. We should
figure out if we want this to be opt-in or opt-out. Currently it's
neither and always enabled if peerpad is loaded via `peerpad.net`.

Relevant people have received an invitation to the rollbar project as
well.

License: MIT
Signed-off-by: Victor Bjelkholm <git@victor.earth>

* chore: update package-lock.json

License: MIT
Signed-off-by: Victor Bjelkholm <git@victor.earth>