feat: remove home-page, jump directly to new pad
This commit removes a lot of code! Mainly, it replaces the home-page
with just a quick call to generate a new pad, and redirects the user
there directly.

The pad itself now features a introduction document with some short
information about PeerPad. I tried to get the markdown source to show up
in the editor as well, but didn't have any luck, and is also outside the
scope of this change in reality.

This commit also includes the removal of the warning banner in the top.

Mathpads are still in the codebase, but no way in the UI to generate
anymore.

License: MIT
Signed-off-by: Victor Bjelkholm <git@victor.earth>